+++ Block/Menu/Edit/Tab/Nodes.php
@@ -8,11 +8,15 @@ use Magento\Cms\Model\Wysiwyg\Config;
 use Magento\Framework\Registry;
 use Snowdog\Menu\Api\NodeRepositoryInterface;
 use Snowdog\Menu\Controller\Adminhtml\Menu\Edit;
+use Snowdog\Menu\Model\Menu\Node\Image\File as ImageFile;
 use Snowdog\Menu\Model\NodeTypeProvider;
 use Snowdog\Menu\Model\VueProvider;
 
 class Nodes extends Template implements TabInterface
 {
+    const IMAGE_UPLOAD_URL = 'snowmenu/node/uploadimage';
+    const IMAGE_DELETE_URL = 'snowmenu/node/deleteimage';
+
     protected $_template = 'menu/nodes.phtml';
 
     /**
@@ -26,6 +30,10 @@ class Nodes extends Template implements TabInterface
     private $nodeRepository;
 
     /**
+     * @var ImageFile
+     */
+    private $imageFile;
+    /**
      * @var NodeTypeProvider
      */
     private $nodeTypeProvider;
@@ -38,6 +46,7 @@ class Nodes extends Template implements TabInterface
     public function __construct(
         Template\Context $context,
         NodeRepositoryInterface $nodeRepository,
+        ImageFile $imageFile,
         NodeTypeProvider $nodeTypeProvider,
         Registry $registry,
         VueProvider $vueProvider,
@@ -47,6 +56,7 @@ class Nodes extends Template implements TabInterface
         $this->registry = $registry;
         $this->nodeRepository = $nodeRepository;
         $this->nodeTypeProvider = $nodeTypeProvider;
+        $this->imageFile = $imageFile;
         $this->vueProvider = $vueProvider;
     }
 
@@ -118,6 +128,30 @@ class Nodes extends Template implements TabInterface
         return false;
     }
 
+    /**
+     * @return string
+     */
+    public function getImageUploadUrl()
+    {
+        return $this->getUrl(self::IMAGE_UPLOAD_URL);
+    }
+
+    /**
+     * @return string
+     */
+    public function getImageDeleteUrl()
+    {
+        return $this->getUrl(self::IMAGE_DELETE_URL);
+    }
+
+    /**
+     * @return string
+     */
+    public function getImageUploadFileId()
+    {
+        return $this->imageFile->getUploadFileId();
+    }
+
     private function renderNodeList($level, $parent, $data)
     {
         if ($parent === null) {
@@ -133,6 +167,7 @@ class Nodes extends Template implements TabInterface
         foreach ($nodes as $node) {
             $menu[] = [
                 'is_active' => $node->getIsActive(),
+                'is_stored' => true,
                 'type' => $node->getType(),
                 'content' => $node->getContent(),
                 'classes' => $node->getClasses(),
@@ -141,9 +176,10 @@ class Nodes extends Template implements TabInterface
                 'submenu_template' => $node->getSubmenuTemplate(),
                 'id' => $node->getId(),
                 'title' => $node->getTitle(),
-                'columns' => $this->renderNodeList($level + 1, $node->getId(), $data)
-                    ? $this->renderNodeList($level + 1, $node->getId(), $data)
-                    : []
+                'image' => $node->getImage(),
+                'image_url' => $node->getImage() ? $this->imageFile->getUrl($node->getImage()) : null,
+                'image_alt_text' => $node->getImageAltText(),
+                'columns' => $this->renderNodeList($level + 1, $node->getId(), $data) ?: []
             ];
         }
         return $menu;