+++ Controller/Adminhtml/Menu/Save.php
@@ -3,18 +3,20 @@
 namespace Snowdog\Menu\Controller\Adminhtml\Menu;
 
 use Magento\Backend\App\Action;
-use Magento\Catalog\Model\ProductRepository;
 use Magento\Framework\Api\FilterBuilderFactory;
 use Magento\Framework\Api\Search\FilterGroupBuilderFactory;
 use Magento\Framework\Api\SearchCriteriaBuilderFactory;
 use Magento\Framework\App\ObjectManager;
 use Magento\Framework\App\ResponseInterface;
-use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Exception\ValidatorException;
 use Snowdog\Menu\Api\Data\MenuInterface;
 use Snowdog\Menu\Api\MenuRepositoryInterface;
 use Snowdog\Menu\Api\NodeRepositoryInterface;
 use Snowdog\Menu\Model\Menu\NodeFactory;
+use Snowdog\Menu\Model\Menu\Node\Image\File as NodeImageFile;
+use Snowdog\Menu\Model\Menu\Node\Image\Node as ImageNode;
 use Snowdog\Menu\Model\MenuFactory;
+use Snowdog\Menu\Model\Menu\Node\Validator as NodeValidator;
 use Snowdog\Menu\Service\MenuHydrator;
 
 class Save extends Action
@@ -42,8 +44,14 @@ class Save extends Action
     /** @var MenuFactory */
     private $menuFactory;
 
-    /** @var ProductRepository */
-    private $productRepository;
+    /** @var NodeValidator */
+    private $nodeValidator;
+
+    /** @var NodeImageFile */
+    private $nodeImageFile;
+
+    /** @var ImageNode */
+    private $imageNode;
 
     /** @var MenuHydrator */
     private $hydrator;
@@ -57,7 +65,9 @@ class Save extends Action
      * @param SearchCriteriaBuilderFactory $searchCriteriaBuilderFactory
      * @param NodeFactory $nodeFactory
      * @param MenuFactory $menuFactory
-     * @param ProductRepository $productRepository
+     * @param NodeValidator $nodeValidator
+     * @param NodeImageFile $nodeImageFile
+     * @param ImageNode $imageNode
      * @param MenuHydrator|null $hydrator
      */
     public function __construct(
@@ -69,7 +79,9 @@ class Save extends Action
         SearchCriteriaBuilderFactory $searchCriteriaBuilderFactory,
         NodeFactory $nodeFactory,
         MenuFactory $menuFactory,
-        ProductRepository $productRepository,
+        NodeValidator $nodeValidator,
+        NodeImageFile $nodeImageFile,
+        ImageNode $imageNode,
         MenuHydrator $hydrator = null
     ) {
         parent::__construct($context);
@@ -80,7 +92,9 @@ class Save extends Action
         $this->searchCriteriaBuilderFactory = $searchCriteriaBuilderFactory;
         $this->nodeFactory = $nodeFactory;
         $this->menuFactory = $menuFactory;
-        $this->productRepository = $productRepository;
+        $this->nodeValidator = $nodeValidator;
+        $this->nodeImageFile = $nodeImageFile;
+        $this->imageNode = $imageNode;
         // Backwards compatible class loader
         $this->hydrator = $hydrator ?? ObjectManager::getInstance()->get(MenuHydrator::class);
     }
@@ -115,14 +129,23 @@ class Save extends Action
         $nodes = $nodes ? json_decode($nodes, true) : [];
         $nodes = $this->_convertTree($nodes, '#');
         $nodeMap = [];
+        $invalidNodes = [];
 
         foreach ($nodes as $node) {
             $nodeId = $node['id'];
+
+            if (!$this->validateNode($node)) {
+                $invalidNodes[$nodeId] = $node;
+            }
 
             if (isset($existingNodes[$nodeId])) {
                 unset($nodesToDelete[$nodeId]);
                 $nodeMap[$nodeId] = $existingNodes[$nodeId];
             } else {
+                if (isset($invalidNodes[$nodeId])) {
+                    continue;
+                }
+
                 $nodeObject = $this->nodeFactory->create();
                 $nodeObject->setMenuId($menu->getMenuId());
                 $nodeObject = $this->nodeRepository->save($nodeObject);
@@ -130,14 +153,21 @@ class Save extends Action
             }
         }
 
-        foreach (array_keys($nodesToDelete) as $nodeId) {
+        $nodesToDeleteIds = array_keys($nodesToDelete);
+        $nodesToDeleteImages = $this->imageNode->getNodeListImages($nodesToDeleteIds);
+
+        foreach ($nodesToDeleteIds as $nodeId) {
             $this->nodeRepository->deleteById($nodeId);
+
+            if (isset($nodesToDeleteImages[$nodeId])) {
+                $this->nodeImageFile->delete($nodesToDeleteImages[$nodeId]);
+            }
         }
 
         $path = ['#' => 0];
 
         foreach ($nodes as $node) {
-            if ($node['type'] == 'product' && !$this->validateProductNode($node)) {
+            if (isset($invalidNodes[$node['id']])) {
                 continue;
             }
 
@@ -191,6 +221,13 @@ class Save extends Action
             $nodeObject->setLevel((string) $level);
             $nodeObject->setPosition((string) $position);
 
+            if ($nodeObject->getImage() && empty($node['image'])) {
+                $this->nodeImageFile->delete($nodeObject->getImage());
+            }
+
+            $nodeObject->setImage($node['image'] ?? null);
+            $nodeObject->setImageAltText($node['image_alt_text'] ?? null);
+
             $this->nodeRepository->save($nodeObject);
 
             $path[$node['id']] = 0;
@@ -242,22 +279,6 @@ class Save extends Action
     }
 
     /**
-     * @param array $node
-     * @return bool
-     */
-    private function validateProductNode(array $node)
-    {
-        try {
-            $this->productRepository->getById($node['content']);
-        } catch (NoSuchEntityException $e) {
-            $this->messageManager->addErrorMessage(__('Product does not exist'));
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
      * Returns menu model based on the Request (requested with `id` or fresh instance)
      *
      * @return MenuInterface
@@ -271,5 +292,18 @@ class Save extends Action
         }
 
         return $this->menuFactory->create();
+    }
+
+    private function validateNode(array $node): bool
+    {
+        try {
+            $this->nodeValidator->validate($node);
+            $result = true;
+        } catch (ValidatorException $exception) {
+            $this->messageManager->addErrorMessage($exception->getMessage());
+            $result = false;
+        }
+
+        return $result;
     }
 }