+++ view/adminhtml/web/js/nodes.js
@@ -2,14 +2,16 @@ define([
     "Vue",
     "Vddl",
     "vue-select",
+    "vue-treeselect",
     "vue!Snowdog_Menu/vue/app",
-    "vue!Snowdog_Menu/vue/nested-list",
-    "vue!Snowdog_Menu/vue/menu-type",
-    "vue!Snowdog_Menu/vue/field-type/simple",
-    "vue!Snowdog_Menu/vue/field-type/checkbox",
     "vue!Snowdog_Menu/vue/field-type/autocomplete",
-    "vue!Snowdog_Menu/vue/field-type/template-list"
-], function(Vue, Vddl, vueSelect) {
+    "vue!Snowdog_Menu/vue/field-type/checkbox",
+    "vue!Snowdog_Menu/vue/field-type/image-upload",
+    "vue!Snowdog_Menu/vue/field-type/simple",
+    "vue!Snowdog_Menu/vue/field-type/template-list",
+    "vue!Snowdog_Menu/vue/menu-type",
+    "vue!Snowdog_Menu/vue/nested-list"
+], function(Vue, Vddl, vueSelect, vueTreeselect) {
 
     return function(config) {
         var dependencies = [];
@@ -22,6 +24,7 @@ define([
             Vue.use(Vddl);
 
             Vue.component('v-select', vueSelect.VueSelect);
+            Vue.component('treeselect', vueTreeselect.Treeselect);
             var app = new Vue({
                 el  : config.el || "#snowdog_menu",
                 data: config.data