+++ view/adminhtml/web/vue/menu-type/category-child.vue
@@ -1,18 +1,27 @@
 <template>
-    <div class="admin__field field">
-        <auto-complete
-            :label="config.translation.categoryChild"
-            :description="config.translation.categoryId"
-            :item="item"
-            :options="config.fieldData.category_child.snowMenuAutoCompleteField.options"
-            :config="config"
-        />
-    </div>
+    <auto-complete
+        :is-tree="true"
+        :label="config.translation.categoryChild"
+        :description="config.translation.categoryId"
+        :item="item"
+        :item-key="'content'"
+        :options="config.fieldData.category_child.snowMenuAutoCompleteField.options"
+        :config="config"
+    />
 </template>
 <script>
     define(["Vue"], function(Vue) {
-        Vue.component("category_child", {
-            props: ['config', 'item'],
+        Vue.component('category_child', {
+            props: {
+                config: {
+                    type: Object,
+                    required: true
+                },
+                item: {
+                    type: Object,
+                    required: true
+                }
+            },
             template: template
         });
     });